# api.yaml
openapi: 3.0.1
info:
  title: Chatic app
  version: 1.0.0
servers:
  - url: http://labvm-42-22.itmo-lab.cosm-lab.science:8080
    description: Common gateway
paths:
...
  /persons/{username}:
    delete:
      summary: Delete user (admin and user itself)
      tags:
        - person-controller
      operationId: deletePerson
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
...
  /personal-chats/{username}:
    get:
      summary: Get info about personal chat with user
      tags:
        - personal-chat-controller
      operationId: getPersonalChat
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonalChatResponseDto'
    post:
      summary: Create personal chat with user
      tags:
        - personal-chat-controller
      operationId: addPersonalChat
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonalChatResponseDto'
  /personal-chats:
    get:
      summary: Get all personal chats (pageable)
      tags:
        - personal-chat-controller
      operationId: getPersonalChats
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseDtoPersonalChatResponseDto'
...
security:
  - jwtAuth : []
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: jwt
  schemas:
...
    PersonalChatResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        person1Id:
          type: string
          format: uuid
        person2Id:
          type: string
          format: uuid
    PageResponseDtoPersonalChatResponseDto:
      type: object
      properties:
        page:
          type: array
          items:
            $ref: '#/components/schemas/PersonalChatResponseDto'
        hasNext:
          type: boolean
        currentPage:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32
...